Moved prime 1 cosmetics test file to fit the main source directory tree, and added a few tests for hue_rotate
